backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
redis==5.0.1
celery==5.3.4

# Development and testing dependencies
black==25.1.0
isort==6.0.1
flake8==7.0.0
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==5.0.0
coverage==7.4.1
